requirements.txt
Django==4.2.16
djangorestframework==3.14.0
mysqlclient==2.2.4